agents/requirements.txt
# ============================================================
# AutoGenius — Agent Module Requirements
# Project 1: CrewAI + LangChain + Groq Multi-Agent System
# ============================================================

# --- Core Agent Framework ---
crewai==0.80.0                  # Multi-agent orchestration (roles, tasks, handoffs)
crewai-tools==0.17.0            # CrewAI built-in tools

# --- LLM Orchestration ---
langchain==0.3.13               # LLM chaining, prompt templates, memory
langchain-groq==0.2.3           # LangChain x Groq integration
langchain-community==0.3.13     # Community tools (web search, loaders, etc.)
langchain-core==0.3.28          # Core LangChain primitives

# --- Groq Client ---
groq==0.13.1                    # Official Groq Python SDK

# --- Output & Formatting ---
pydantic==2.10.4                # Data validation & structured output schemas
rich==13.9.4                    # Beautiful terminal output (agent logs)
tabulate==0.9.0                 # Table formatting for reports

# --- Utilities ---
python-dotenv==1.0.1            # Load .env for API keys
uuid==1.30                      # Unique run IDs
tiktoken==0.8.0                 # Token counting

# --- Async Support ---
asyncio==3.4.3                  # Async agent execution
nest-asyncio==1.6.0             # Nested async loops (Jupyter/scripts)
